Add real-data benchmark and move bench scripts to bench/
- New bench/benchmark_real.py: tests all VLMs and LLMs against real
  annotated video clips (violence detection, eating, animal threat)
- bench/data/annotations.json: clip manifest with ROI timestamps,
  app questions, and positive/negative keyword lists
- bench/data/videos/: directory structure for clips (.gitkeep tracked,
  .mp4/.webm/.mkv gitignored)
- bench/results/: default output directory for JSON results
- Moved benchmark_detectors/llms/vlms.py from scripts/ to bench/
- justfile + pyproject.toml: lint/format coverage extended to bench/
- .gitignore: ignore video files under bench/data/videos/
